Added makefiles and corrected warnings

+++ Mixed Fractions/main.cpp
@@ -9,11 +9,11 @@ int main(int argc, char **argv)
   {
     numbers.push_back(num);
   }
-  for (int i = 0; i < numbers.size() - 2; i += 2)
+  for (size_t i = 0; i < numbers.size() - 2; i += 2)
   {
     nom = numbers.at(i);
     denom = numbers.at(i + 1);
-    std::cout << (int) nom / denom << " " << nom % denom << " / " << denom << 
+    std::cout << (int) nom / denom << " " << nom % denom << " / " << denom <<
      "\n";
   }
   return 0;
